put processing in main function
